debug: debug for script, need rollback later
