+++ Cargo.toml
@@ -28,6 +28,7 @@ bitcoin-script-stack = { git = "https://github.com/FairgateLabs/rust-bitcoin-scr
 rand = "0.8.5"
 rand_chacha = "0.3.1"
 ethereum-types = "0.9"
+once_cell = "1.19.0"
 
 [dev-dependencies]
 num-bigint = { version = "0.4.4", features = ["rand"] }
