+++ src/groth16/mod.rs
@@ -1,3 +1,4 @@
+pub mod constants;
 pub mod offchain_checker;
 pub mod pairing;
 mod test;
